requirements.txt
python-gitlab==8.0.0